Remove legacy symbol programs.
