+++ include/mbgl/shaders/shader_manifest.hpp
@@ -35,8 +35,4 @@
 #include <mbgl/shaders/gl/drawable_wide_vector.hpp>
 #include <mbgl/shaders/gl/prelude.hpp>
 #include <mbgl/shaders/gl/clipping_mask.hpp>
-#include <mbgl/shaders/gl/symbol_icon.hpp>
-#include <mbgl/shaders/gl/symbol_sdf_text.hpp>
-#include <mbgl/shaders/gl/symbol_sdf_icon.hpp>
-#include <mbgl/shaders/gl/symbol_text_and_icon.hpp>
 #endif
