--- src/mbgl/programs/symbol_sdf_text_program.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-// Alias
-#include <mbgl/programs/symbol_program.hpp>
