--- src/mbgl/programs/symbol_text_and_icon_program.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-// Alias
-#include <mbgl/programs/symbol_program.hpp>
